+++ noodles-bam/src/bai/index/reference_sequence/builder.rs
@@ -60,7 +60,7 @@ impl Builder {
     }
 
     fn update_bins(&mut self, record: &Record, chunk: Chunk) -> io::Result<()> {
-        let bin_id = match (record.alignment_start(), record.alignment_end()) {
+        let bin_id = match (record.position(), record.alignment_end()) {
             (Some(start), Some(end)) => region_to_bin(start, end).map(u32::from)?,
             _ => u32::from(UNMAPPED_BIN),
         };