+++ noodles-cram/src/reader/record.rs
@@ -107,9 +107,7 @@ where
             self.read_mapped_read(&mut record, cram_bit_flags, read_length)?;
         }
 
-        self.prev_alignment_start = record
-            .alignment_start()
-            .and_then(|start| Position::new(i32::from(start) as usize));
+        self.prev_alignment_start = record.alignment_start();
 
         Ok(record)
     }