+++ noodles-cram/src/record/convert.rs
@@ -28,6 +28,8 @@ impl Record {
         }
 
         if let Some(alignment_start) = self.alignment_start() {
+            let alignment_start =
+                sam::record::Position::try_from(usize::from(alignment_start) as i32).unwrap();
             builder = builder.set_position(alignment_start);
         }
 
@@ -48,6 +50,8 @@ impl Record {
         }
 
         if let Some(mate_alignment_start) = self.mate_alignment_start() {
+            let mate_alignment_start =
+                sam::record::Position::try_from(usize::from(mate_alignment_start) as i32).unwrap();
             builder = builder.set_mate_position(mate_alignment_start);
         }